% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check the data is ready to be analyzed}
\usage{
check_data(data)
}
\arguments{
\item{data}{Data frame to be passed to run_pca() \code{data} argument.}
}
\description{
Checks if the data is appropriate for subsequent analysis.
}
\details{
Checks for missing variables, missing values in exposure variables, ...
}
\examples{
library(dplyr)
data(expanse)
check_data(expanse)

expanse_invalid_column <- expanse \%>\%
  rename(no2_2010_annual = no2)
try(check_data(expanse_invalid_column))

expanse_missing <- expanse \%>\%
  mutate(ndvi = ifelse(ndvi < 1, NA, ndvi))
try(check_data(expanse_missing))
}
\author{
Sergio Olmos
}
