% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expanse}
\alias{expanse}
\title{Example data set for Paper 1 in EXPANSE WP3}
\format{
A data frame with 1000 rows and 16 variables:
\describe{
\item{id}{Individual ID}
\item{no2}{NO2}
\item{bc}{BC}
\item{pm25}{PM2.5}
\item{o3}{Ozone}
\item{ndvi}{NDVI}
\item{imperv}{Impervious surfaces density}
\item{dist_water}{Distance to water}
\item{temp_cold_mean}{Mean temperature during the cold season}
\item{temp_cold_sd}{Temperature standard deviation during cold season}
\item{temp_warm_mean}{Mean temperature during warm season}
\item{temp_warm_sd}{Temperature standard deviation during warm season}
\item{ses}{SES index (numeric)}
\item{ses_std}{Standardized SES index (numeric)}
\item{ses_cat}{SES category (factor)}
\item{pop_density}{Population density}
Note that the exposure variables in your data set must have these exact
same names as the corresponding exposures in this example data set for
it to work with the rexpanse1 functions.
}
}
\source{
Simulated.
}
\usage{
expanse
}
\description{
A data set with simulated exposure variables and covariates.
}
\keyword{datasets}
