% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ses.R
\name{fit_ses}
\alias{fit_ses}
\title{Fit SES models}
\usage{
fit_ses(data, pc_var, ses_var, other_vars)
}
\value{

}
\description{
Fit models assessing relationship between SES and principal components.
}
\author{
Sergio Olmos
}
