% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pca_domains.R
\name{get_pca_domains}
\alias{get_pca_domains}
\title{Compute PCA for all domains}
\usage{
get_pca_domains(data, summary = TRUE)
}
\arguments{
\item{data}{Data frame containing all 11 exposure variables across the three
different domains.}

\item{summary}{Logical vector of length one indicating whether only
a summary of the PCA should be returned. See Details.}
}
\value{
A list containing a summary of the PCA, the actual principal component
values together with the variables used, and a table identifying the
correction factor of each principal component for standardizing
the interpretation across cohorts.
}
\description{
Computes principal component analysis for all domains.
}
\details{
The domains are overall, air pollution, built environment and temperature.
}
\examples{
data(expanse)
get_pca_domains(expanse)
}
\author{
Sergio Olmos
}
