% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quantiles.R
\name{get_quantiles}
\alias{get_quantiles}
\title{Compute exposure quantiles}
\usage{
get_quantiles(data)
}
\arguments{
\item{data}{Data frame containing the exposures of interest. See Details.}
}
\value{
A matrix with quantiles (rows) for each exposure (columns).
}
\description{
Computes exposure quantiles for making exposure boxplots.
}
\details{
The data frame passed to \code{data} should contain the eleven exposure
variables with the agreed variable names:
\itemize{
\item no2
\item pm25
\item bc
\item o3
\item ndvi
\item imperv
\item dist_water
\item temp_warm_mean
\item temp_warm_sd
\item temp_cold_mean
\item temp_cold_mean
}
}
\examples{
data(expanse)
get_quantiles(expanse)
}
\author{
Sergio Olmos
}
