% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quantiles_by_domain.R
\name{get_quantiles_by_domain}
\alias{get_quantiles_by_domain}
\title{Compute exposure quantiles}
\usage{
get_quantiles_by_domain(data, domain)
}
\arguments{
\item{data}{Data frame containing the exposures of interest. See Details.}
}
\value{
A data frame with quantiles and domain (rows) for each exposure (columns).
}
\description{
Computes exposure quantiles for making exposure boxplots, adding domain in this cases SES variable.
}
\details{
The data frame passed to \code{data} should contain the eleven exposure
variables with the agreed variable names:
\itemize{
\item NO22010
\item PM252010
\item BC2010
\item O32010w
\item ndvi_2019
\item ImpSurf_2015
\item Dist_water
\item temp_2010_warm
\item temp_2010_warm_sd
\item temp_2010_cold
\item temp_2010_cold_sd
\item ses_cat_indv
\item ses_cat_area
}
}
\examples{
data(expanse)
get_quantiles(expanse, domain = ses_cat_var)
}
\author{
Fabian Coloma
}
