% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_pca_summary.R
\name{pivot_pca_summary}
\alias{pivot_pca_summary}
\title{Pivot PCA summary into long format}
\usage{
pivot_pca_summary(pca)
}
\arguments{
\item{pca}{List as returned by get_pca().}
}
\value{
A tibble with PCA summary in long format.
}
\description{
Reshapes the PCA summary information into a long format.
}
\details{
Each row corresponds to a variable in each component.
Note that the variables pc_sd, pc_var_pop, and pc_cum_var_prop refer to
the principal component and have therefore the same value for all variables
within a principal component.
}
\author{
Sergio Olmos
}
