% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{run_pca}
\alias{run_pca}
\title{Run full analysis pipeline}
\usage{
run_pca(data, ses_var, other_vars, cohort_name, write = TRUE, output_dir = ".")
}
\value{

}
\description{
Generates all results for Paper 1.
}
\author{
Sergio Olmos
}
