% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_1_by_domain.R
\name{table1_extract}
\alias{table1_extract}
\title{Compute Table 1}
\usage{
table1_extract(data, ses_var = NULL, age_var = NULL)
}
\arguments{
\item{data}{Data frame containing posible SES (max. 2)
and Age, it can be at base line or real.}

\item{ses_var}{Vector of length max. 2 containing the name of SES.
this vector can be NULL.}

\item{age_var}{Name of age variable}
}
\value{
A list containing a summary of Age variable
Number of individuos
Proportion of Indv. by SES.
}
\description{
Computes table 1 - summary of principal data
}
\details{
The domains are SES and age.
}
\examples{
data(expanse)
table1_extract(data, ses_var = "ses_cat_indv", age_var = "age_random")
table1_extract(data, ses_var = c("ses_cat_indv", "ses_cat_area"), age_var = "age_random")
table1_extract(data, ses_var = NULL, age_var = NULL)
}
\author{
Fabian Coloma
}
