% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check the data is ready to be analyzed}
\usage{
check_data(data)
}
\arguments{
\item{data}{Data frame to be passed to run_pca() \code{data} argument.}
}
\description{
Checks if the data is appropriate for subsequent analysis.
}
\details{
Checks for missing variables, missing values in exposure variables, ...
}
\examples{
data(exposures)
check_data(exposures)
}
\author{
Sergio Olmos
}
