% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quantiles.R
\name{get_quantiles}
\alias{get_quantiles}
\title{Compute exposure quantiles}
\usage{
get_quantiles(data)
}
\arguments{
\item{data}{Data frame containing the exposures of interest. See Details.}
}
\value{
A matrix with quantiles (rows) for each exposure (columns).
}
\description{
Computes exposure quantiles for making exposure boxplots.
}
\details{
The data frame passed to \code{data} should contain the eleven exposure
variables with the agreed variable names:
\itemize{
\item NO22010
\item PM252010
\item BC2010
\item O32010w
\item ndvi_2019
\item ImpSurf_2015
\item Dist_water
\item temp_2010_warm
\item temp_2010_warm_sd
\item temp_2010_cold
\item temp_2010_cold_sd
}
}
\examples{
data(expanse)
get_quantiles(expanse)
}
\author{
Sergio Olmos
}
