% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pca.R
\name{run_pca}
\alias{run_pca}
\title{Run full analysis pipeline}
\usage{
run_pca(
  data,
  ses_var = NULL,
  age_var = "age",
  other_vars,
  cohort_name,
  write = TRUE,
  output_dir = "."
)
}
\arguments{
\item{data}{Data frame containing all exposures}

\item{ses_var}{Charcater vector of length one or two indicating the name of
the SES variable (individual, and by region) in \code{data} which will be used as a predictor when looking
at the association with the principal components.}

\item{age_var}{character indicating the name of Age at baseline variable as numeric.}

\item{other_vars}{Character vector indicating the names of any other
variables in \code{data} that need to be included as covariates in the
models assessing the association between the principal components
and SES.}

\item{cohort_name}{Character vector of length one indicating the name of
the cohort.}

\item{write}{Logical vector of length one indicating whether the output
should be written to disk.}

\item{output_dir}{Character vector of length one indicating the path where
the output file should be written if \code{write = TRUE}. Defaults to current
working directory.}
}
\value{
A list containing the following elements:
}
\description{
Generates all results for Paper 1.
}
\examples{
data(expanse)
results_dir <- tempdir()
run_pca(
  data = expanse,
  ses_var = "ses_cat_***",
  age_var = "age",
  cohort_name = "catalonia",
  write = TRUE,
  output_dir = results_dir
)

results_dir <- tempdir()
run_pca(
  data = expanse,
  ses_var = c("ses_cat_inv", "ses_cat_area"),
  age_var = "age",
  cohort_name = "catalonia",
  write = TRUE,
  output_dir = results_dir
)
}
\author{
Sergio Olmos
}
